Config project

+++ GameScopeBackend/app/Http/Controllers/ComparisonController.php
@@ -2,8 +2,9 @@
 
 namespace App\Http\Controllers;
 
-use Illuminate\Http\Request;
 use App\Models\Comparison;
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class ComparisonController extends Controller
 {
@@ -14,37 +15,35 @@ class ComparisonController extends Controller
 
     public function store(Request $request)
     {
-        $validatedData = $request->validate([
+        $request->validate([
             'game_id1' => 'required|exists:video_games,id',
             'game_id2' => 'required|exists:video_games,id',
-            'details' => 'required|string',
+            'details' => 'nullable|string'
         ]);
 
-        $comparison = Comparison::create($validatedData);
+        $comparison = Comparison::create(array_merge($request->all(), ['user_id' => Auth::id()]));
 
         return response()->json($comparison, 201);
     }
 
-    public function show(Comparison $comparison)
+    public function show($id)
     {
-        return $comparison;
+        return Comparison::findOrFail($id);
     }
 
-    public function update(Request $request, Comparison $comparison)
+    public function update(Request $request, $id)
     {
-        $validatedData = $request->validate([
-            'game_id1' => 'sometimes|required|exists:video_games,id',
-            'game_id2' => 'sometimes|required|exists:video_games,id',
-            'details' => 'sometimes|required|string',
-        ]);
+        $comparison = Comparison::findOrFail($id);
+        $this->authorize('update', $comparison);
+        $comparison->update($request->all());
 
-        $comparison->update($validatedData);
-
-        return response()->json($comparison);
+        return response()->json($comparison, 200);
     }
 
-    public function destroy(Comparison $comparison)
+    public function destroy($id)
     {
+        $comparison = Comparison::findOrFail($id);
+        $this->authorize('delete', $comparison);
         $comparison->delete();
 
         return response()->json(null, 204);
